something

+++ src/university/courses/Transcript.java
@@ -1,21 +1,16 @@
 package university.courses;
 
 
-public class Transcript
-{
+public class Transcript {
 	public Marks marks;
-	
-
 	public Transcript(){
 		super();
 	}
 
 
 	
-	public int calculateGPA() {
-		// TODO implement me
-		return 0;	
+	public double calculateGPA() {
+		return 0;
 	}
-	
 }
